set the secondView have bug ,because upData

+++ mall/interface/second/onSquareView/secondListViewController.h
@@ -10,6 +10,6 @@
 
 @interface secondListViewController : UIViewController
 
-@property (assign, nonatomic) NSInteger  gc_ID;
+@property (assign, nonatomic) int  gc_ID;
 
 @end
